# backend/src/docs/gen/swagger.yaml
basePath: /api
definitions:
...
  github_com_moasq_backend_app_organizations_app_services.AddMemberResponse:
    properties:
      email:
        type: string
      invite_sent:
        type: boolean
      member_id:
        type: string
      name:
        type: string
      org_id:
        type: string
      role_slug:
        type: string
    type: object
  github_com_moasq_backend_app_organizations_app_services.BootstrapOrganizationRequest:
    properties:
      org_display_name:
        description: Organization details
        type: string
      owner_email:
        description: Owner member details
        type: string
      owner_name:
        type: string
    required:
    - org_display_name
    - owner_email
    - owner_name
    type: object
  github_com_moasq_backend_app_organizations_app_services.BootstrapOrganizationResponse:
    properties:
      display_name:
        type: string
      invite_sent:
        type: boolean
      magic_link_sent:
        type: boolean
      org_slug:
        type: string
      organization_id:
        type: string
      owner_email:
        type: string
      owner_member_id:
        type: string
      owner_name:
        type: string
    type: object
  github_com_moasq_backend_app_organizations_app_services.ListMembersResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/github_com_moasq_backend_app_organizations_app_services.MemberInfo'
        type: array
      total:
        type: integer
    type: object
  github_com_moasq_backend_app_organizations_app_services.MemberInfo:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      member_id:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_moasq_backend_app_organizations_app_services.ProfileOrganization:
    properties:
      name:
        type: string
      organization_id:
        type: string
      slug:
        type: string
      status:
        type: string
    type: object
  github_com_moasq_backend_app_organizations_app_services.ProfileResponse:
    properties:
      account_id:
        description: Internal account details
        type: integer
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      member_id:
        description: Auth provider member details
        type: string
      name:
        type: string
      organization:
        allOf:
        - $ref: '#/definitions/github_com_moasq_backend_app_organizations_app_services.ProfileOrganization'
        description: Organization details
      permissions:
        items:
          type: string
        type: array
      roles:
        items:
          type: string
        type: array
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for B2B SaaS Starter.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: B2B SaaS Starter API
  version: "1.0"
paths:
...
  /auth/check-email:
    get:
      consumes:
      - application/json
      description: Checks if an email exists in any organization. Returns 200 OK (empty
        response) if exists, 404 Not Found if doesn't exist. This is a public endpoint
        used during login flow.
      parameters:
      - description: Email address to check
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email exists
        "400":
          description: Invalid email format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Email not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Check if email exists
      tags:
      - auth
  /auth/members:
    get:
      consumes:
      - application/json
      description: Retrieves all members of the current organization. Restricted to
        admin role only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_moasq_backend_app_organizations_app_services.ListMembersResponse'
        "400":
          description: Missing organization context
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Insufficient permissions - admin role required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to list members
          schema:
            additionalProperties: true
            type: object
      summary: List organization members
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: 'Adds a new member to an existing organization with a specified
        role. Organization ID is automatically extracted from JWT token. Member receives
        a magic link invite email for passwordless authentication. Request body: {"email":
        "user@example.com", "name": "Full Name", "role_slug": "member"}'
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Member email address
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: Member full name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Role slug (defaults to 'member')
        in: body
        name: role_slug
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_moasq_backend_app_organizations_app_services.AddMemberResponse'
        "400":
          description: Invalid request payload or missing organization context
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to add member
          schema:
            additionalProperties: true
            type: object
      summary: Add member to organization
      tags:
      - auth
  /auth/members/{member_id}:
    delete:
      consumes:
      - application/json
      description: Removes a member from the organization (deletes from both Stytch
        and internal database). Only admins can delete members.
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Member ID to delete
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Member deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid member ID or missing organization context
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Insufficient permissions - admin role required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Member not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete member
          schema:
            additionalProperties: true
            type: object
      summary: Delete organization member
      tags:
      - auth
  /auth/profile/me:
    get:
      consumes:
      - application/json
      description: Retrieves comprehensive profile information for the currently authenticated
        user, including member details, organization info, and account status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_moasq_backend_app_organizations_app_services.ProfileResponse'
        "400":
          description: Missing required context (organization or claims)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve profile
          schema:
            additionalProperties: true
            type: object
      summary: Get current user profile
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new organization in Stytch with an initial admin member.
        The admin receives a magic link invite email to complete passwordless onboarding.
        Organization slug is auto-generated from the organization name.
      parameters:
      - description: Organization bootstrap request (passwordless - no password required)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_moasq_backend_app_organizations_app_services.BootstrapOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_moasq_backend_app_organizations_app_services.BootstrapOrganizationResponse'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to bootstrap organization
          schema:
            additionalProperties: true
            type: object
      summary: Bootstrap organization
      tags:
      - auth
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"